- Added default path for log storage - Program can now run in background
