+++ values.py
@@ -1,0 +1,5 @@
+from pynput import keyboard
+
+COMBINATION = {keyboard.KeyCode.from_char('\x03')}
+DEFAULT_PATH = 'C:\\Clipboard-log\\'
+FILE_PATH = 'C:\\Clipboard-log\\'
